♻️  [REFACTOR] 대기실 및 랭킹 페이지 구조 개선
- 대기실 페이지를 새로운 구조로 리팩토링하여 코드 가독성 향상
- 랭킹 페이지를 ranking 폴더로 이동하고, 관련 컴포넌트로 분리
- 기존 WaitingRoomPage 및 RankingPage 삭제 후 새로운 구조로 통합
- 대기실 및 랭킹 관련 컴포넌트 추가 및 개선
