+++ src/App.tsx
@@ -13,13 +13,13 @@ const LoginPage = lazy(() => import("./pages/login/LoginPage"));
 const SignupPage = lazy(() => import("./pages/signup/SignupPage"));
 const ProfileSetupPage = lazy(() => import("./pages/setup-profile/ProfileSetupPage"));
 const MatchingPage = lazy(() => import("./pages/matching/MatchingPage"));
-const WaitingRoomPage = lazy(() => import("./pages/WaitingRoomPage"));
+const WaitingRoomPage = lazy(() => import("./pages/waiting-room/WaitingRoomPage"));
 const ScreenShareSetupPage = lazy(() => import("./pages/ScreenShareSetupPage"));
 const BattlePage = lazy(() => import("./pages/BattlePage"));
 const ResultPage = lazy(() => import("./pages/ResultPage"));
 const TierPromotionPage = lazy(() => import("./pages/TierPromotionPage"));
 const TierDemotionPage = lazy(() => import("./pages/TierDemotionPage"));
-const RankingPage = lazy(() => import("./pages/RankingPage"));
+const RankingPage = lazy(() => import("./pages/ranking/RankingPage"));
 const ProfilePage = lazy(() => import("./pages/profile/ProfilePage"));
 const SettingsPage = lazy(() => import("./pages/settings/SettingsPage"));
 const NotFound = lazy(() => import("./pages/not-found/NotFound"));
