+++ src/pages/waiting-room/components/PlayerCard.tsx
@@ -1,0 +1,74 @@
+import CyberCard from "@/components/CyberCard";
+import CyberButton from "@/components/CyberButton";
+import { User } from "lucide-react";
+import { FC } from "react";
+
+interface Props {
+  name: string;
+  mmr: number;
+  role: "host" | "player";
+  isHostView: boolean;
+  isReady: boolean;
+  onToggleReady?: () => void;
+  onGameStart?: () => void;
+}
+
+const PlayerCard: FC<Props> = ({
+  name,
+  mmr,
+  role,
+  isHostView,
+  isReady,
+  onToggleReady,
+  onGameStart,
+}) => (
+  <CyberCard glowing={role === "host"}>
+    <div className="text-center space-y-4">
+      <div
+        className={`w-20 h-20 ${role === "host" ? "bg-gradient-to-r from-cyber-blue to-cyber-purple" : "bg-gradient-to-r from-gray-600 to-gray-700"} rounded-full mx-auto flex items-center justify-center`}
+      >
+        <User className="h-10 w-10 text-white" />
+      </div>
+      <div>
+        <h3 className="text-xl font-bold text-white">{name}</h3>
+        <p
+          className={
+            role === "host" ? "text-cyber-blue font-semibold" : "text-gray-400"
+          }
+        >
+          {role === "host" ? "방장" : "플레이어"}
+        </p>
+        <p className="text-sm text-gray-400">MMR: {mmr}</p>
+      </div>
+      {role === "host" ? (
+        isHostView ? (
+          <CyberButton
+            onClick={onGameStart}
+            disabled={!isReady}
+            className="w-full"
+          >
+            게임 시작
+          </CyberButton>
+        ) : (
+          <div className="text-green-400 font-semibold">준비 완료</div>
+        )
+      ) : isHostView ? (
+        <div
+          className={`font-semibold ${isReady ? "text-green-400" : "text-yellow-400"}`}
+        >
+          {isReady ? "준비 완료" : "준비 중..."}
+        </div>
+      ) : (
+        <CyberButton
+          onClick={onToggleReady}
+          variant={isReady ? "secondary" : "primary"}
+          className="w-full"
+        >
+          {isReady ? "준비 취소" : "준비"}
+        </CyberButton>
+      )}
+    </div>
+  </CyberCard>
+);
+
+export default PlayerCard;
